refactor: 스크롤 위치(차지 공간 등) 개선
