+++ src/components/common/StagingArea/StagingAreaTaskContainer.tsx
@@ -1,0 +1,180 @@
+import styled from '@emotion/styled';
+
+import BtnTask from '@/components/common/BtnTask/BtnTask';
+import ScrollGradient from '@/components/common/ScrollGradient';
+import StagingAreaSetting from '@/components/common/StagingArea/StagingAreaSetting';
+import { TaskType } from '@/types/tasks/taskType';
+
+function StagingAreaTaskContainer() {
+	const dummyTaskList: TaskType[] = [
+		{
+			id: 0,
+			name: '바보~',
+			deadLine: {
+				date: '2024-06-30',
+				time: '12:30',
+			},
+			hasDescription: false,
+			status: '진행중',
+		},
+		{
+			id: 1,
+			name: '넛수레',
+			deadLine: {
+				date: '2024-06-30',
+				time: '12:30',
+			},
+			hasDescription: true,
+			status: '지연',
+		},
+		{
+			id: 2,
+			name: '콘하스',
+			deadLine: {
+				date: '2024-06-30',
+				time: '12:30',
+			},
+			hasDescription: true,
+			status: '완료',
+		},
+		{
+			id: 3,
+			name: '김지원',
+			deadLine: {
+				date: '2024-06-30',
+				time: '12:30',
+			},
+			hasDescription: true,
+			status: '미완료',
+		},
+		{
+			id: 0,
+			name: '바보~',
+			deadLine: {
+				date: '2024-06-30',
+				time: '12:30',
+			},
+			hasDescription: false,
+			status: '진행중',
+		},
+		{
+			id: 1,
+			name: '넛수레',
+			deadLine: {
+				date: '2024-06-30',
+				time: '12:30',
+			},
+			hasDescription: true,
+			status: '지연',
+		},
+		{
+			id: 2,
+			name: '콘하스',
+			deadLine: {
+				date: '2024-06-30',
+				time: '12:30',
+			},
+			hasDescription: true,
+			status: '완료',
+		},
+		{
+			id: 3,
+			name: '김지원',
+			deadLine: {
+				date: '2024-06-30',
+				time: '12:30',
+			},
+			hasDescription: true,
+			status: '미완료',
+		},
+		{
+			id: 0,
+			name: '바보~',
+			deadLine: {
+				date: '2024-06-30',
+				time: '12:30',
+			},
+			hasDescription: false,
+			status: '진행중',
+		},
+		{
+			id: 1,
+			name: '넛수레',
+			deadLine: {
+				date: '2024-06-30',
+				time: '12:30',
+			},
+			hasDescription: true,
+			status: '지연',
+		},
+		{
+			id: 2,
+			name: '콘하스',
+			deadLine: {
+				date: '2024-06-30',
+				time: '12:30',
+			},
+			hasDescription: true,
+			status: '완료',
+		},
+		{
+			id: 3,
+			name: '김지원',
+			deadLine: {
+				date: '2024-06-30',
+				time: '12:30',
+			},
+			hasDescription: true,
+			status: '미완료',
+		},
+	];
+
+	return (
+		<StagingAreaTaskContainerLayout>
+			<StagingAreaSetting />
+			<BtnTaskContainer>
+				{dummyTaskList.map((task) => (
+					<BtnTask
+						key={task.id + task.name}
+						btnType="staging"
+						hasDescription={task.hasDescription}
+						id={task.id}
+						name={task.name}
+						status={task.status}
+						deadLine={task.deadLine}
+					/>
+				))}
+				<ScrollGradient />
+			</BtnTaskContainer>
+		</StagingAreaTaskContainerLayout>
+	);
+}
+
+export default StagingAreaTaskContainer;
+
+const StagingAreaTaskContainerLayout = styled.div`
+	display: flex;
+	flex-direction: column;
+	gap: 1.3rem;
+	align-items: flex-start;
+	align-self: stretch;
+`;
+const BtnTaskContainer = styled.div`
+	position: relative;
+	display: flex;
+	flex-direction: column;
+	gap: 1rem;
+	width: 100%;
+	height: 56.7rem;
+	overflow: auto;
+	overflow-y: scroll;
+
+	::-webkit-scrollbar {
+		width: 0.6rem;
+	}
+
+	::-webkit-scrollbar-thumb {
+		background-color: ${({ theme }) => theme.palette.Grey.Grey6};
+		border-radius: 3px;
+	}
+`;
